+++ src/main/java/com/webapps2022/entity/CurrencyTransaction.java
@@ -5,7 +5,7 @@
 package com.webapps2022.entity;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
+
 import java.time.Instant;
 import java.util.Objects;
 import javax.persistence.Entity;
@@ -43,10 +43,10 @@ public class CurrencyTransaction implements Serializable {
     SystemUser fromSystemUser;
 
     @NotNull
-    BigDecimal currencyCountTo;
+    Float currencyCountTo;
 
     @NotNull
-    BigDecimal currencyCountFrom;
+    Float currencyCountFrom;
 
     @NotNull
     CurrencyEnum currencyTypeTo;
@@ -63,7 +63,7 @@ public class CurrencyTransaction implements Serializable {
     public CurrencyTransaction() {
     }
 
-    public CurrencyTransaction(SystemUser toSystemUser, SystemUser fromSystemUser, BigDecimal currencyCountTo, BigDecimal currencyCountFrom, CurrencyEnum currencyTypeTo, CurrencyEnum currencyTypeFrom, Instant timeStamp, boolean completed) {
+    public CurrencyTransaction(SystemUser toSystemUser, SystemUser fromSystemUser, Float currencyCountTo, Float currencyCountFrom, CurrencyEnum currencyTypeTo, CurrencyEnum currencyTypeFrom, Instant timeStamp, boolean completed) {
         this.toSystemUser = toSystemUser;
         this.fromSystemUser = fromSystemUser;
         this.currencyCountTo = currencyCountTo;
@@ -98,19 +98,19 @@ public class CurrencyTransaction implements Serializable {
         this.fromSystemUser = fromSystemUser;
     }
 
-    public BigDecimal getCurrencyCountTo() {
+    public Float getCurrencyCountTo() {
         return currencyCountTo;
     }
 
-    public void setCurrencyCountTo(BigDecimal currencyCountTo) {
+    public void setCurrencyCountTo(Float currencyCountTo) {
         this.currencyCountTo = currencyCountTo;
     }
 
-    public BigDecimal getCurrencyCountFrom() {
+    public Float getCurrencyCountFrom() {
         return currencyCountFrom;
     }
 
-    public void setCurrencyCountFrom(BigDecimal currencyCountFrom) {
+    public void setCurrencyCountFrom(Float currencyCountFrom) {
         this.currencyCountFrom = currencyCountFrom;
     }
